# docs/swagger/internal-admin.yaml
openapi: 3.1.0
info:
  title: Officely - Internal Admin API
  version: 1.2.1
servers: 
  - url: https://officelybackend-bbgmg9hnc8cff9hp.polandcentral-01.azurewebsites.net/admin
  - url: https://officely-dev.noscorp.pl/admin
  - url: http://localhost:8080/admin
components: 
  securitySchemes: 
    user_auth:
      name: User auth
      type: http
      in: header
      scheme: "Bearer"
  schemas:
    Link:
      type: object
      required:
        - href
      properties:
        href:
          type: string
    PageSize:
      type: integer
      minimum: 1
      maximum: 50
      examples: [25]
...
    OfficeResource:
      type: object
      required: 
        - id
        - ownerId
        - name
        - description
        - openingHours
        - address
        - coordinates
        - photoUrls
        - contactEmail
        - contactPhone
        - paymentAccountNumber
        - paymentReceiverName
        - _links
      properties: 
        id:
          type: string
        ownerId:
          type: string
        name:
          type: string
        description:
          type: string
        openingHours:
          type: string
        address:
          type: string
        coordinates:
          type: object
          required:
            - lat
            - lon
          properties:
            lat:
              type: number
            lon:
              type: number
        photoUrls:
          type: array
          items:
            type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        published:
          type: boolean
          default: false
        paymentAccountNumber:
          type: string
        paymentReceiverName:
          type: string
        _links:
          type: object
          required:
            - self
            - update
            - offers
            - items
            - members
          properties:
            self:
              $ref: '#/components/schemas/Link'
            update:
              $ref: '#/components/schemas/Link'
            offers:
              $ref: '#/components/schemas/Link'
            items:
              $ref: '#/components/schemas/Link'
            members:
              $ref: '#/components/schemas/Link'
    OfficeOfferResource:
      type: object
      required: 
        - id
        - officeId
        - name
        - pricePerDay
        - pricePerDayCurrency
        - freeCancellationHours
        - paymentHours
        - properties
        - _links
      properties: 
        id:
          type: string
        officeId:
          type: string
        name:
          type: string
        publicName:
          type: string
        pricePerDay:
          type: integer
        pricePerDayCurrency:
          type: string
        freeCancellationHours:
          type: integer
        paymentHours:
          type: integer
          description: The payment must be made x hours after the booking was made
        properties:
          type: object
          additionalProperties: 
            anyOf: 
              - type: string
              - type: boolean
              - type: integer
            
        _links:
          type: object
          required:
            - self
            - office
          properties:
            self:
              $ref: '#/components/schemas/Link'
            office:
              $ref: '#/components/schemas/Link'
    OfficeItemResource:
      type: object
      required: 
        - id
        - officeId
        - name
        - floor
        - room
        - offerId
        - type
        - _links
      properties: 
        id:
          type: string
        officeId:
          type: string
        name:
          type: string
        floor:
          type: string
        room:
          type: string
        offerId:
          type: string
        type:
          type: string
          enum:
            - SHARED
            - INDIVIDUAL
        capacity:
          type: integer
          minimum: 1
          description: Required iff type is shared, otherwise must not be present
        _links:
          type: object
          required:
            - self
            - office
            - offer
          properties:
            self:
              $ref: '#/components/schemas/Link'
            office:
              $ref: '#/components/schemas/Link'
            offer:
              $ref: '#/components/schemas/Link'
...
paths:
...
  '/offices/{officeId}':
    get:
      tags: [ Offices ]
      summary: "View the details of an office (if the user is allowed to view it)"
      security: 
        - user_auth: []
      parameters: 
        - in: path
          name: officeId
          required: true
          schema: 
            type: string
      responses: 
        "200":
          description: Success
          content:
            application/json: 
              schema: 
                $ref: '#/components/schemas/OfficeResource'
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
    patch:
      tags: [ Offices ]
      summary: Modify an office (only the owner may use this)
      security:
        - user_auth: []
      parameters:
        - in: path
          name: officeId
          required: true
          schema: 
            type: string
      requestBody: 
        required: true
        content: 
          multipart/form-data: 
            schema: 
              type: object
              required:
                - office
              properties:
                office:
                  type: object
                  properties:
                    name:
                      type: string
                      minLength: 1
                      maxLength: 64
                    description:
                      type: string
                      minLength: 1
                      maxLength: 4096
                    openingHours:
                      type: string
                      minLength: 1
                      maxLength: 1024
                    address:
                      type: string
                      minLength: 1
                      maxLength: 256
                    contactEmail:
                      type: string
                      minLength: 1
                      maxLength: 256
                      format: email
                    contactPhone:
                      type: string
                      pattern: "^\\+[1-9]\\d{1,14}$"
                    paymentAccountNumber:
                      type: string
                      minLength: 1
                      maxLength: 34
                      description: In the international IBAN format
                    paymentReceiverName:
                      type: string
                      minLength: 1
                      maxLength: 64
                    published:
                      type: boolean
                    images:
                      type: array
                      description: Ordered list of images to be used, images not in this list will be removed, current refers to the active images and added refers to images included in this request
                      minItems: 1
                      maxItems: 10
                      items:
                        type: string
                      examples:
                        - - "current[0]"
                          - "added[0]"
                          - "current[2]"
                addedImages:
                  type: array
                  maxItems: 10
                  items:
                    type: string
                    format: binary
            encoding: 
              office:
                contentType: application/json
      responses:
        "204":
          description: Success
        "400":
          description: "Validation error"
          content: 
            application/json: 
              schema: 
                type: object
                required:
                  - errors
                properties:
                  errors:
                    type: array
                    minItems: 1
                    items:
                      type: object
                      oneOf: 
                        - type: object
                          required:
                            - type
                            - field
                            - message
                          properties:
                            type:
                              type: string
                              const: ValidationError
                            field:
                              type: string
                              enum:
                                - name
                                - description
                                - openingHours
                                - address
                                - contactEmail
                                - contactPhone
                                - paymentAccountNumber
                                - paymentReceiverName
                                - published
                                - images
                                - addedImages
                            message:
                              type: string
                              description: A human readable message
                              examples:
                                - Value must be a boolean
                        - type: object
                          required:
                            - type
                          properties:
                            type:
                              type: string
                              const: InvalidBodyError
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
    
  # ====== OFFERS =======
  '/offices/{officeId}/offers':
    get:
      tags: [ Offers ]
      summary: "Get a list of all offers belonging to the office"
      security: 
        - user_auth: []
      parameters: 
        - in: path
          name: officeId
          required: true
          schema: 
            type: string
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: sortField
          schema:
            type: string
            examples:
              - firstName
        - in: query
          name: sortDirection
          schema:
            type: string
            enum:
              - asc
              - desc
        - name: pageSize
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/PageSize'
        - name: pageToken
          in: query
          schema:
            type: integer
      responses: 
        "200":
          description: Success
          content:
            application/json: 
              schema: 
                type: object
                required:
                  - results
                  - _pagination
                  - _links
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/OfficeOfferResource'
                      
                  _pagination:
                    type: object
                    required: 
                      - currentPage
                      - lastPage
                      - pageSize
                    properties:
                      currentPage:
                        type: integer
                      lastPage:
                        type: integer
                      pageSize:
                        type: integer
                  _links:
                    type: object
                    required:
                      - self
                      - first
                      - last
                    properties:
                      self:
                        $ref: '#/components/schemas/Link'
                      next:
                        $ref: '#/components/schemas/Link'
                      prev:
                        $ref: '#/components/schemas/Link'
                      first:
                        $ref: '#/components/schemas/Link'
                      last:
                        $ref: '#/components/schemas/Link'
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
    post:
      tags: [ Offers ]
      summary: Create an offer
      security:
        - user_auth: []
      parameters: 
        - in: path
          name: officeId
          required: true
          schema: 
            type: string
      requestBody: 
        required: true
        content: 
          application/json: 
            schema: 
              type: object
              required: 
                - name
                - pricePerDay
                - pricePerDayCurrency
                - paymentHours
                - properties
              properties:
                sourceId:
                  type: string
                  description: "If present, will be take over all items linked to the source offer, so it will make all items linked to the source offer link to this offer instead"
                name:
                  type: string
                  minLength: 1
                  maxLength: 64
                publicName:
                  type: string
                  minLength: 1
                  maxLength: 64
                pricePerDay:
                  type: integer
                  minimum: 1
                pricePerDayCurrency:
                  type: string
                  minLength: 1
                  maxLength: 3
                  examples:
                    - PLN
                freeCancellationHours:
                  type: integer
                  minimum: 0
                  maximum: 744 # 24*31
                paymentHours:
                  type: integer
                  minimum: 0
                  maximum: 744 # 24*31
                  description: The payment must be made x hours after the booking was made
                properties:
                  type: object
                  additionalProperties:
                    type: string
      responses:
        "201":
          description: Created
          headers: 
            Location:
              example: /offices/2056/offers/2056
              description: URL to the newly created offer (for example /offices/2056/offers/2056)
              schema: 
                type: string
          content: 
            application/json: 
              schema: 
                type: object
                required:
                  - id
                properties:
                  id:
                    type: string
        "400":
          description: "Validation error"
          content: 
            application/json: 
              schema: 
                type: object
                required:
                  - errors
                properties:
                  errors:
                    type: array
                    minItems: 1
                    items:
                      type: object
                      oneOf: 
                        - type: object
                          required:
                            - type
                            - field
                            - message
                          properties:
                            type:
                              type: string
                              const: ValidationError
                            field:
                              type: string
                              enum:
                                - name
                                - publicName
                                - pricePerDay
                                - pricePerDayCurrency
                                - freeCancellationHours
                                - paymentHours
                                - properties
                                - images
                            message:
                              type: string
                              description: A human readable message
                              examples:
                                - Name is required
                        - type: object
                          required:
                            - type
                          properties:
                            type:
                              type: string
                              const: InvalidBodyError
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
  '/offices/{officeId}/offers/{offerId}':
    get:
      tags: [ Offers ]
      summary: "View the details of an offer"
      security: 
        - user_auth: []
      parameters: 
        - in: path
          name: officeId
          required: true
          schema: 
            type: string
        - in: path
          name: offerId
          required: true
          schema: 
            type: string
      responses: 
        "200":
          description: Success
          content:
            application/json: 
              schema: 
                $ref: '#/components/schemas/OfficeOfferResource'
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
    patch:
      tags: [ Offers ]
      summary: Modify an offer
      security:
        - user_auth: []
      parameters:
        - in: path
          name: officeId
          required: true
          schema: 
            type: string
        - in: path
          name: offerId
          required: true
          schema: 
            type: string
      requestBody: 
        required: true
        content: 
          application/json: 
            schema: 
              type: object
              properties:
                available:
                  type: boolean
      responses:
        "204":
          description: Success
        "400":
          description: "Validation error"
          content: 
            application/json: 
              schema: 
                type: object
                required:
                  - errors
                properties:
                  errors:
                    type: array
                    minItems: 1
                    items:
                      type: object
                      oneOf: 
                        - type: object
                          required:
                            - type
                            - field
                            - message
                          properties:
                            type:
                              type: string
                              const: ValidationError
                            field:
                              type: string
                              enum:
                                - available
                            message:
                              type: string
                              description: A human readable message
                              examples:
                                - Value must be a boolean
                        - type: object
                          required:
                            - type
                          properties:
                            type:
                              type: string
                              const: InvalidBodyError
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found


  # ====== Items =======
...
  '/offices/{officeId}/items/{itemId}':
    get:
      tags: [ Items ]
      summary: "View the details of an item"
      security: 
        - user_auth: []
      parameters: 
        - in: path
          name: officeId
          required: true
          schema: 
            type: string
        - in: path
          name: itemId
          required: true
          schema: 
            type: string
      responses: 
        "200":
          description: Success
          content:
            application/json: 
              schema: 
                $ref: '#/components/schemas/OfficeItemResource'
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
    patch:
      tags: [ Items ]
      summary: Modify an item
      security:
        - user_auth: []
      parameters:
        - in: path
          name: officeId
          required: true
          schema: 
            type: string
        - in: path
          name: itemId
          required: true
          schema: 
            type: string
      requestBody: 
        required: true
        content: 
          application/json: 
            schema: 
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 64
                floor:
                  type: string
                  minLength: 1
                  maxLength: 8
                room:
                  type: string
                  minLength: 1
                  maxLength: 16
                offerId:
                  type: string
                available:
                  type: boolean
                capacity:
                  type: integer
                  minimum: 1
                  description: Only applicable to type SHARED
      responses:
        "204":
          description: Success
        "400":
          description: "Validation error"
          content: 
            application/json: 
              schema: 
                type: object
                required:
                  - errors
                properties:
                  errors:
                    type: array
                    minItems: 1
                    items:
                      type: object
                      oneOf: 
                        - type: object
                          required:
                            - type
                            - field
                            - message
                          properties:
                            type:
                              type: string
                              const: ValidationError
                            field:
                              type: string
                              enum:
                                - name
                                - floor
                                - room
                                - offerId
                                - available
                                - capacity
                            message:
                              type: string
                              description: A human readable message
                              examples:
                                - Value must be a boolean
                        - type: object
                          required:
                            - type
                          properties:
                            type:
                              type: string
                              const: InvalidBodyError
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found


  # ====== Members =======
...